+++ permutect/test/tools/test_refine_permutect_model.py
@@ -4,13 +4,13 @@ from argparse import Namespace
 from tensorboard.backend.event_processing.event_accumulator import EventAccumulator
 from permutect import constants
 from permutect.architecture.permutect_model import load_model
-from permutect.tools import train_artifact_model
+from permutect.tools import refine_permutect_model
 
 
-def test_train_artifact_model():
+def test_refine_permutect_model():
     # Inputs
     training_data_tarfile = '/Users/davidben/mutect3/permutect/integration-tests/singular-10-Mb/preprocessed-dataset.tar'
-    base_model = '/Users/davidben/mutect3/permutect/integration-tests/singular-10-Mb/base-model.pt'
+    pretrained_model = '/Users/davidben/mutect3/permutect/integration-tests/singular-10-Mb/base-model.pt'
 
     # Outputs
     saved_model = tempfile.NamedTemporaryFile()
@@ -28,7 +28,7 @@ def test_train_artifact_model():
 
     # Training data inputs
     setattr(train_model_args, constants.TRAIN_TAR_NAME, training_data_tarfile)
-    setattr(train_model_args, constants.SAVED_MODEL_NAME, base_model)
+    setattr(train_model_args, constants.SAVED_MODEL_NAME, pretrained_model)
 
     # training hyperparameters
     setattr(train_model_args, constants.BATCH_SIZE_NAME, 64)
@@ -43,7 +43,7 @@ def test_train_artifact_model():
     setattr(train_model_args, constants.OUTPUT_NAME, saved_model.name)
     setattr(train_model_args, constants.TENSORBOARD_DIR_NAME, training_tensorboard_dir.name)
 
-    train_artifact_model.main_without_parsing(train_model_args)
+    refine_permutect_model.main_without_parsing(train_model_args)
 
     events = EventAccumulator(training_tensorboard_dir.name)
     events.Reload()
