+++ permutect/test/tools/test_train_permutect_model.py
@@ -5,7 +5,7 @@ from tensorboard.backend.event_processing.event_accumulator import EventAccumula
 
 from permutect import constants
 from permutect.architecture.permutect_model import load_model
-from permutect.tools import train_base_model
+from permutect.tools import train_permutect_model
 
 
 def test_train_permutect_model():
@@ -47,7 +47,7 @@ def test_train_permutect_model():
     setattr(train_model_args, constants.OUTPUT_NAME, saved_model.name)
     setattr(train_model_args, constants.TENSORBOARD_DIR_NAME, training_tensorboard_dir.name)
 
-    train_base_model.main_without_parsing(train_model_args)
+    train_permutect_model.main_without_parsing(train_model_args)
 
     events = EventAccumulator(training_tensorboard_dir.name)
     events.Reload()
