+++ permutect/tools/prune_dataset.py
@@ -19,7 +19,7 @@ from permutect import constants, utils
 from permutect.data.artifact_dataset import ArtifactDataset
 from permutect.parameters import add_training_params_to_parser, TrainingParameters
 from permutect.data.base_dataset import BaseDataset
-from permutect.tools.train_artifact_model import parse_training_params
+from permutect.tools.refine_permutect_model import parse_training_params
 from permutect.utils import Label
 
 NUM_FOLDS = 3
@@ -204,7 +204,7 @@ def parse_arguments():
     # input / output
     parser.add_argument('--' + constants.TRAIN_TAR_NAME, type=str, required=True,
                         help='tarfile of training/validation datasets produced by preprocess_dataset.py')
-    parser.add_argument('--' + constants.SAVED_MODEL_NAME, type=str, help='Base model from train_base_model.py')
+    parser.add_argument('--' + constants.SAVED_MODEL_NAME, type=str, help='Base model from train_permutect_model.py')
     parser.add_argument('--' + constants.OUTPUT_NAME, type=str, required=True, help='path to pruned dataset file')
     parser.add_argument('--' + constants.TENSORBOARD_DIR_NAME, type=str, default='tensorboard', required=False,
                         help='path to output tensorboard directory')
